Refactor visitor interface and internals
Traversal (depth-first / breadth-first) and order (pre / post) are now
members of mCc_ast_visitor. We therefore no longer need to forward them
along with the visitor struct.

This greatly improves readability as most calls to internal functions
and macros now fit on one line.

The internal macros now examine the visitor struct's members and invoke
the callback with the correct arguments.

This also allows us to drop the variadic macros.